Maybe the most horrible thing I've ever done in C
Instead of just sending commands to the VNC, also send them to serial. To achieve this, I've made Tee, which behaves a bit like tee on UNIX. Except that it's programmed with 50% jank and 50% bodge, and gives a warning every time it's used for a const char.

Such is life.